Pin setup-sp to v1.2.1
Since rumblefrog/setup-sp#7 is resolved, we can safely use the
latest available version of that in the workflows.  To avoid
potential breakages in the future, we'll leave it to users of this
template to update the workflow version as they see fit.

The template currently has no guarantees that it'll behave the same
with spcomp 1.12 (1.11+?) because of compiler changes.